Add warning to floodGame.js if you're logged in

+++ cheats/mobileGui.js
@@ -431,44 +431,45 @@
                 {
                     name: "Flood Game",
                     description: "Floods a game with a number of fake accounts",
-                    run: async function () {
+                    run: function () {
                         let i = document.createElement('iframe');
                         document.body.append(i);
                         window.prompt = i.contentWindow.prompt.bind(window);
+                        window.alert = i.contentWindow.alert.bind(window);
                         i.remove();
-        
-                        const id = prompt("Game ID:");
-                        const name = prompt("Name:");
-                        const amount = parseInt(prompt("Amount:"));
         
                         let cache = Object.values(webpackJsonp.push([[], { ['']: (_, a, b) => { a.cache = b.c }, }, [['']]]).cache);
                         const axios = cache.find((x) => x.exports?.a?.get).exports.a;
                         const firebase = cache.find(x => x.exports?.a?.initializeApp).exports.a;
-                        const blooks = Object.keys(Object.values(cache).find(x => x.exports.a?.Black).exports.a);
+                        const blooks = Object.keys(cache.find(x => x.exports.a?.Black).exports.a);
         
-                        for (let i = 1; i <= amount; i++) {
-                            (async () => {
-                                let ign = `${name}${String.fromCharCode(96 + i)}`;
-                                const { data: { success, fbToken, fbShardURL } } = await axios.put("https://fb.blooket.com/c/firebase/join", { id, name: ign });
-                                if (!success) return;
-                                const liveApp = firebase.initializeApp({
-                                    apiKey: "AIzaSyCA-cTOnX19f6LFnDVVsHXya3k6ByP_MnU",
-                                    authDomain: "blooket-2020.firebaseapp.com",
-                                    projectId: "blooket-2020",
-                                    storageBucket: "blooket-2020.appspot.com",
-                                    messagingSenderId: "741533559105",
-                                    appId: "1:741533559105:web:b8cbb10e6123f2913519c0",
-                                    measurementId: "G-S3H5NGN10Z",
-                                    databaseURL: fbShardURL
-                                }, ign);
-                                const auth = firebase.auth(liveApp);
-                                await auth.setPersistence(firebase.auth.Auth.Persistence.NONE).catch(console.error);
-                                await auth.signInWithCustomToken(fbToken).catch(console.error);
-                                await liveApp.database().ref(`${id}/c/${ign}`).set({ b: blooks[Math.floor(Math.random() * blooks.length)] });
-                                liveApp.delete();
-                            })();
-                            await new Promise(r => setTimeout(r, 100));
-                        }
+                        cache.find(x => x.exports?.a?.me).exports.a.me({}).then(async data => {
+                            if (data.name) return alert("You are logged in, and using this script will suspend your account. Please log out if you wish to use this.");
+                            const id = prompt("Game ID:"), name = prompt("Name:"), amount = parseInt(prompt("Amount:"));
+                            for (let i = 1; i <= amount; i++) {
+                                (async () => {
+                                    let ign = `${name}${String.fromCharCode(96 + i)}`;
+                                    const { data: { success, fbToken, fbShardURL } } = await axios.put("https://fb.blooket.com/c/firebase/join", { id, name: ign });
+                                    if (!success) return;
+                                    const liveApp = firebase.initializeApp({
+                                        apiKey: "AIzaSyCA-cTOnX19f6LFnDVVsHXya3k6ByP_MnU",
+                                        authDomain: "blooket-2020.firebaseapp.com",
+                                        projectId: "blooket-2020",
+                                        storageBucket: "blooket-2020.appspot.com",
+                                        messagingSenderId: "741533559105",
+                                        appId: "1:741533559105:web:b8cbb10e6123f2913519c0",
+                                        measurementId: "G-S3H5NGN10Z",
+                                        databaseURL: fbShardURL
+                                    }, ign);
+                                    const auth = firebase.auth(liveApp);
+                                    await auth.setPersistence(firebase.auth.Auth.Persistence.NONE).catch(console.error);
+                                    await auth.signInWithCustomToken(fbToken).catch(console.error);
+                                    await liveApp.database().ref(`${id}/c/${ign}`).set({ b: blooks[Math.floor(Math.random() * blooks.length)] });
+                                    liveApp.delete();
+                                })();
+                                await new Promise(r => setTimeout(r, 100));
+                            }
+                        });
                     }
                 },
                 {
@@ -2697,7 +2698,7 @@
         }
         let iframe = document.querySelector("iframe");
         const [_, time, error] = decode.match(/LastUpdated: (.+?); ErrorMessage: "(.+?)"/);
-        if (parseInt(time) <= 1700445441874 || iframe.contentWindow.confirm(error)) cheat();
+        if (parseInt(time) <= 1700494615673 || iframe.contentWindow.confirm(error)) cheat();
     }
     img.onerror = img.onabort = () => (img.src = null, cheat());
 })();
